+++ launch-or-update-local.sh
@@ -8,13 +8,15 @@ function displayHelp {
             Set repo to latest dev versions (defaults to true)
           -u, --update [true|false]
             Update Helm Charts (defaults to true)
+          -r, --useRemote [true|false]
+            Use remote repo (defaults to true)
           "
   exit
 }
 
 # Option strings
-SHORT="hl:u:"
-LONG="help,latest:,update:"
+SHORT="hl:u:r:"
+LONG="help,latest:,update:,useRemote:"
 
 # read the options
 OPTS=$(getopt --options $SHORT --long $LONG --name "$0" -- "$@")
@@ -39,9 +41,15 @@ while true ; do
       fi
       shift 2
       ;;
-      -u | --update )
+    -u | --update )
       if [ "$2" = "false" ]; then 
           update="false"
+      fi
+      shift 2
+      ;;
+    -r | --useRemote )
+      if [ "$2" = "false" ]; then 
+          useRemote="false"
       fi
       shift 2
       ;;
@@ -61,18 +69,20 @@ if [[ "$context" != "docker-desktop" ]]; then
   exit 1
 fi
 
-if [[ "$latest" != "false" ]]; then
-  echo -e "Getting Latest Chart Versions... \n"
-  ./update-chart-versions.sh -v development
-else
-  echo -e "Getting Master Chart Versions... \n"
-  ./update-chart-versions.sh -v master
-fi 
+if [[ "$useRemote" != "false" ]]; then
+  if [[ "$latest" != "false" ]]; then
+    echo -e "Getting Latest Chart Versions... \n"
+    ./update-chart-versions.sh -v development
+  else
+    echo -e "Getting Master Chart Versions... \n"
+    ./update-chart-versions.sh -v master
+  fi 
 
-if [[ "$update" != "false" ]]; then
-    echo -e "\n Updating Dependencies... \n"
-    rm $chart/charts/*.tgz
-    helm dependency update $chart
+  if [[ "$update" != "false" ]]; then
+      echo -e "\n Updating Dependencies... \n"
+      rm $chart/charts/*.tgz
+      helm dependency update $chart
+  fi
 fi
 
 if [[ "$OSTYPE" == "darwin"* ]]; then
